Fixed several bugs relating to truncation.
Truncating to a smaller size the requries the block size to be reduced is not working. The file metadata reflects the new size however the number of blocks remains unchanged.

To Replicate with block size 4096:
touch mnt/test
truncate --size 5000
truncate --size 500

View database their are still two blocks when only one is necessary.